+++ pkg/resources/block_storage/resource_volume.go
@@ -2,6 +2,7 @@ package block_storage
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"strings"
 
@@ -333,6 +334,13 @@ func (r *ResourceVolume) Read(ctx context.Context, req resource.ReadRequest, res
 		id,
 	)
 	if err != nil {
+
+		if errors.Is(err, exoscale.ErrNotFound) {
+			// Resource doesn't exist anymore, signaling the core to remove it from the state.
+			resp.State.RemoveResource(ctx)
+			return
+		}
+
 		resp.Diagnostics.AddError(
 			"unable to get storage volume",
 			err.Error(),